move SeriesLoader and SeriesLocator into series.py
